Modify: 리뷰, 담기, 찜하기 에러핸들링

+++ services/addList.service.js
@@ -1,38 +1,43 @@
 const addListDao = require('../models/addList.dao');
 
 const addBookshelf = async (user_id, books_id) => {
-  try {
+  const check = await addListDao.checkAddedBookshelf(user_id, books_id);
+  if (check[0]) {
+    throw { statusCode: 400, message: 'ALREADY BOOKSHELF ADDED' };
+  } else {
     return await addListDao.addBookshelf(user_id, books_id);
-  } catch (err) {
-    console.log(err);
-    res.status(err.statusCode).json({ message: err.message });
   }
 };
 
 const removeBookshelf = async (user_id, books_id) => {
-  try {
+  const check = await addListDao.checkAddedBookshelf(user_id, books_id);
+  if (!check[0]) {
+    throw { statusCode: 400, message: 'BOOKSHELF IS NOT EXIST' };
+  } else if (check[0].users_id !== user_id) {
+    throw { statusCode: 400, message: 'ONLY OWNER CAN DELETE' };
+  } else {
     return await addListDao.removeBookshelf(user_id, books_id);
-  } catch (err) {
-    console.log(err);
-    res.status(err.statusCode).json({ message: err.message });
   }
 };
 
 const addFavorite = async (user_id, books_id) => {
-  try {
+  const check = await addListDao.checkAddedFavorite(user_id, books_id);
+  console.log('check =', check);
+  if (check[0]) {
+    throw { statusCode: 400, message: 'ALREADY FAVORITE ADDED' };
+  } else {
     return await addListDao.addFavorite(user_id, books_id);
-  } catch (err) {
-    console.log(err);
-    res.status(err.statusCode).json({ message: err.message });
   }
 };
 
 const removeFavorite = async (user_id, books_id) => {
-  try {
+  const check = await addListDao.checkAddedFavorite(user_id, books_id);
+  if (!check[0]) {
+    throw { statusCode: 400, message: 'FAVORITE IS NOT EXIST' };
+  } else if (check[0].users_id !== user_id) {
+    throw { statusCode: 400, message: 'ONLY OWNER CAN DELETE' };
+  } else {
     return await addListDao.removeFavorite(user_id, books_id);
-  } catch (err) {
-    console.log(err);
-    res.status(err.statusCode).json({ message: err.message });
   }
 };
